restructure project, update go version
