+++ README.md
@@ -2,16 +2,19 @@
 
 A project to allow multiple users to collaborate remotely on crosswords
 
+## Prerequisites
+
+Go v.1.14.2
+
 ## Project Structure
 
 ```
 /bin
-/src
-	/cmd
-		main.go <- our main project file
-	/internal
-		/pkg <- will contain any internal packages (for use only within this project)
-	/pkg <- will contain any non-internal packages (for use within and without of this project)
+/cmd
+	main.go <- our main project file
+/internal
+	/pkg <- will contain any internal packages (for use only within this project)
+/pkg <- will contain any non-internal packages (for use within and without of this project)]
 makefile
 README.md
 ```
@@ -20,12 +23,8 @@ README.md
 
 ### Make commands
 
-* `make clean` - cleans the project
-* `make build` - builds the project into a binary file
-* `make run` - runs the project
+- `make clean` - cleans the project
+- `make build` - builds the project into a binary file
+- `make run` - runs the project
 
 Then navigate to localhost:8080
-
-
-
-
