+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/paulinshek/cryptic-colab
+
+go 1.14
