--- src/internal/pkg/core/grid.go
@@ -1,12 +1,0 @@
-package core
-
-type Direction int
-
-const (
-	Down Direction = iota
-	Across Direction = iota
-)
-
-type Coordinate struct {
-	X, Y int
-}
